rename from dumper to dumpwriter
